+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable user-facing changes to `ticket` are documented here.
 
+## [0.2.1] - 2026-09-24
+
+- Added `ticket check --active` to verify active tickets only.
+- Added TQL ticket-ID prefix and ordered ID filtering.
+
 ## [0.2.0] - 2026-09-23
 
 ### Added
@@ -121,6 +126,7 @@ All notable user-facing changes to `ticket` are documented here.
 
 - Initial release
 
+[0.2.1]: https://github.com/toolsupply/ticket/releases/tag/v0.2.1
 [0.2.0]: https://github.com/toolsupply/ticket/releases/tag/v0.2.0
 [0.1.8]: https://github.com/toolsupply/ticket/releases/tag/v0.1.8
 [0.1.7]: https://github.com/toolsupply/ticket/releases/tag/v0.1.7